feat: eliminate RDS requirement
The database adapter and parameters can be defined in the
node (custom json) if an RDS resource is not defined.
This is expected behaviour for Opsworks users.

BREAKING CHANGE: Sqlite3 is no longer set as the default database
adapter.

In order to use sqlite as the database adapter it must be defined
in the node.