fix: Set the default DB adapter to `sqlite3`
Currently, if no DB adapter is set, both in stack JSON file, and application (no RDS resource
attached), the application crashes with `"There is no supported Db driver for given configuration."`
exception. This is not wanted behavior, as the main goal of this cookbook is to work in
zero-configuration environment. The fix resolves that problem, by simply setting the default DB
adapter to `sqlite3`.