fix: `node['applications']` are back
Unfortunatelly during the setup phase, OpsWorks deploys all applications
in given stack - there is no way to assign some applications to some
layers only. This parameter resolves this issue (and thus - should be
kept) by filtering supported applications in given layer, to only listed
in it. Notice, that with this option enabled, it won't be possible to
deploy to a given layer an application which is not listed in it.

Resolves #55