fix: Bump faraday middleware version
Per https://github.com/lostisland/faraday_middleware/issues/152, bundler
started reporting that 0.11.0 could not be installed due to a checksum
mismatch. To remedy the issue, the faraday middleware team has released
a patch version to allow users to bundle install from scratch again.

I am unsure as to whether or not this project prefers 1.13.17 for
bundler, or always the latest release but I am including the change here
as it is relevant when hitting this issue.