fix: moved all pid-related files from shared/pids to /run/lock
The main reason behind this fix is, that sometimes between server restarts,
PID files remain on the file system, causing `start` scripts to crash.
On AWS /run/lock is a special directory, designed for that, usually with tmpfs
mounted in it. This means, that after each restart this directory gets wiped out.

Fixes #92