Watchtower Updates (rbenv)
This commit updates the repository in order to allow the usage of rbenv instead of ng-ruby. This will allow us to control the version of Ruby with much better control.

Eventually I would like to DRY up some of this code, but it is a slow process since it is trial and error. Due to how chef runs recipes and provided libraries, we are re-initializing rbenv in multiple places, which works, but looks slightly dirty. We are also monkey-patching the Chef::Provider::Deploy#migrate method in order to make that rbenv aware (if available), but still allow the hooks around that migrate command to work properly. That would be nice to get into it’s own file as well, but since it also needs the rbenv stuff, it makes it a bit difficult to simply extract that into it’s own file.